ch(refactor):refactor the working directory - change the working directory of ctrim Finishes
