+++ src/App.js
@@ -4,7 +4,6 @@ import { BrowserRouter,Route,Switch } from 'react-router-dom';
 import HomepageLayout from './components/home';
 import Dashboard from './components/dashboard';
 import login from './components/login';
-import StoreDetail from './components/storeDetail';
 
 class App extends Component {
   render() {
@@ -15,7 +14,6 @@ class App extends Component {
             <Route exact path='/' component={HomepageLayout}/>
             <Route exact path='/dashboard' component={Dashboard}/>
             <Route exact path='/login' component={login}/>
-            <Route exact path='/store/:pk' component={StoreDetail}/>
           </Switch>
         </div>
       </BrowserRouter>
